requirements.txt
django
gunicorn
python-dotenv
psycopg2-binary
whitenoise
djangorestframework
